move mod: move `read_clock()` to mod `shader_clock`
