+++ crates/spirv-std/src/arch.rs
@@ -8,7 +8,6 @@ use crate::Integer;
 use crate::{Scalar, SignedInteger, UnsignedInteger, Vector};
 #[cfg(target_arch = "spirv")]
 use core::arch::asm;
-use glam::UVec2;
 
 mod atomics;
 mod barrier;
@@ -115,51 +114,6 @@ pub unsafe fn vector_insert_dynamic<T: Scalar, V: Vector<T, N>, const N: usize>(
             element = in(reg) &element,
             result = in(reg) &mut result,
         }
-
-        result
-    }
-}
-
-/// Read from the shader clock with either the `Subgroup` or `Device` scope.
-///
-/// See:
-/// <https://htmlpreview.github.io/?https://github.com/KhronosGroup/SPIRV-Registry/blob/master/extensions/KHR/SPV_KHR_shader_clock.html>
-#[spirv_std_macros::gpu_only]
-#[doc(alias = "OpReadClockKHR")]
-pub fn read_clock_khr<const SCOPE: u32>() -> u64 {
-    unsafe {
-        let mut result: u64;
-
-        asm! {
-            "%uint = OpTypeInt 32 0",
-            "%scope = OpConstant %uint {scope}",
-            "{result} = OpReadClockKHR typeof*{result} %scope",
-            result = out(reg) result,
-            scope = const SCOPE,
-        };
-
-        result
-    }
-}
-
-/// Like `read_clock_khr` but returns a vector to avoid requiring the `Int64`
-/// capability. It returns a 'vector of two-components of 32-bit unsigned
-/// integer type with the first component containing the 32 least significant
-/// bits and the second component containing the 32 most significant bits.'
-#[spirv_std_macros::gpu_only]
-#[doc(alias = "OpReadClockKHR")]
-pub fn read_clock_uvec2_khr<const SCOPE: u32>() -> UVec2 {
-    unsafe {
-        let mut result = UVec2::default();
-
-        asm! {
-            "%uint = OpTypeInt 32 0",
-            "%scope = OpConstant %uint {scope}",
-            "%result = OpReadClockKHR typeof*{result} %scope",
-            "OpStore {result} %result",
-            result = in(reg) &mut result,
-            scope = const SCOPE,
-        };
 
         result
     }
